+++ test/src/json_impl_test.dart
@@ -6,9 +6,8 @@
 import 'package:test/test.dart';
 
 // Project imports:
-import 'package:json_response/json_response.dart';
+import 'package:json_response/src/json_impl.dart';
 
-/// This is a test class for [JsonResponse].
 void main() {
   _testFromJsonString();
   _testFromJsonMap();
@@ -36,49 +35,49 @@ void main() {
 
 void _testFromJsonString() {
   test('Test fromJsonString.', () {
-    final json = JsonResponse.fromString(value: '{"test": true}');
+    final json = JsonImpl.fromString(value: '{"test": true}');
     expect(json.getBool(key: 'test'), true);
   });
 }
 
 void _testFromJsonMap() {
   test('Test fromJsonMap.', () {
-    final json = JsonResponse.fromMap(value: {"test": true});
+    final json = JsonImpl.fromMap(value: {"test": true});
     expect(json.getBool(key: 'test'), true);
   });
 }
 
 void _testIsEmptyWhenJsonIsNotEmpty() {
   test('Test isEmpty when JSON is not empty.', () {
-    final json = JsonResponse.fromMap(value: {'test': true});
+    final json = JsonImpl.fromMap(value: {'test': true});
     expect(json.isEmpty, false);
   });
 }
 
 void _testIsEmptyWhenJsonIsEmpty() {
   test('Test isEmpty when JSON is empty.', () {
-    final json = JsonResponse.fromMap(value: {});
+    final json = JsonImpl.fromMap(value: {});
     expect(json.isEmpty, true);
   });
 }
 
 void _testIsNotEmptyWhenJsonIsNotEmpty() {
   test('Test isNotEmpty when JSON is not empty.', () {
-    final json = JsonResponse.fromMap(value: {'test': true});
+    final json = JsonImpl.fromMap(value: {'test': true});
     expect(json.isNotEmpty, true);
   });
 }
 
 void _testIsNotEmptyWhenJsonIsEmpty() {
   test('Test isNotEmpty when JSON is empty.', () {
-    final json = JsonResponse.fromMap(value: {});
+    final json = JsonImpl.fromMap(value: {});
     expect(json.isNotEmpty, false);
   });
 }
 
 void _testContainsKey() {
   test('Test containsKey', () {
-    final json = JsonResponse.fromMap(value: {
+    final json = JsonImpl.fromMap(value: {
       'test': {'test2': true},
       'test3': [
         {'test4': 1}
@@ -95,7 +94,7 @@ void _testContainsKey() {
 
 void _testKeySet() {
   test('Test keySet', () {
-    final json = JsonResponse.fromMap(
+    final json = JsonImpl.fromMap(
       value: {'test1': '', 'test2': '', 'test5': ''},
     );
 
@@ -106,7 +105,7 @@ void _testKeySet() {
 
 void _testGetStringValue() {
   test('Test getStringValue.', () {
-    final json = JsonResponse.fromMap(value: {'test': 'success'});
+    final json = JsonImpl.fromMap(value: {'test': 'success'});
     expect(json.isEmpty, false);
     expect(json.getString(key: 'test'), 'success');
     expect(json.getString(key: 'not_exist'), '');
@@ -117,7 +116,7 @@ void _testGetStringValue() {
 
 void _testGetIntValue() {
   test('Test getIntValue.', () {
-    final json = JsonResponse.fromMap(value: {'test': 1});
+    final json = JsonImpl.fromMap(value: {'test': 1});
     expect(json.isEmpty, false);
     expect(json.getInt(key: 'test'), 1);
     expect(json.getInt(key: 'not_exist'), -1);
@@ -127,7 +126,7 @@ void _testGetIntValue() {
 
 void _testGetDoubleValue() {
   test('Test getDoubleValue.', () {
-    final json = JsonResponse.fromMap(value: {'test': 1.0});
+    final json = JsonImpl.fromMap(value: {'test': 1.0});
     expect(json.isEmpty, false);
     expect(json.getDouble(key: 'test'), 1.0);
     expect(json.getDouble(key: 'not_exist'), -1.0);
@@ -137,7 +136,7 @@ void _testGetDoubleValue() {
 
 void _testGetDoubleValues() {
   test('Test getDoubleValues.', () {
-    final json = JsonResponse.fromMap(value: {
+    final json = JsonImpl.fromMap(value: {
       'test': [1.0, -0.1, 0.0]
     });
     expect(json.isEmpty, false);
@@ -147,7 +146,7 @@ void _testGetDoubleValues() {
 
 void _testGetBoolValue() {
   test('Test getBoolValue.', () {
-    final json = JsonResponse.fromMap(value: {'test': true});
+    final json = JsonImpl.fromMap(value: {'test': true});
     expect(json.isEmpty, false);
     expect(json.getBool(key: 'test'), true);
     expect(json.getBool(key: 'not_exist'), false);
@@ -157,7 +156,7 @@ void _testGetBoolValue() {
 
 void _testGetStringValues() {
   test('Test getStringValues.', () {
-    final json = JsonResponse.fromMap(value: {
+    final json = JsonImpl.fromMap(value: {
       'test': ['example_1', 'example_2']
     });
 
@@ -169,7 +168,7 @@ void _testGetStringValues() {
 
 void _testGetIntValues() {
   test('Test getIntValues.', () {
-    final json = JsonResponse.fromMap(value: {
+    final json = JsonImpl.fromMap(value: {
       'test': [0, 1]
     });
 
@@ -181,10 +180,10 @@ void _testGetIntValues() {
 
 void _testGetJsonFromJsonString() {
   test('Test getJson from JSON String.', () {
-    final json = JsonResponse.fromMap(value: {'test1': '{"test2": true}'});
+    final json = JsonImpl.fromMap(value: {'test1': '{"test2": true}'});
     expect(json.isEmpty, false);
 
-    final childJson = json.getJson(key: 'test1');
+    final childJson = json.get(key: 'test1');
     expect(childJson.isEmpty, false);
     expect(childJson.getBool(key: 'test2'), true);
   });
@@ -192,12 +191,12 @@ void _testGetJsonFromJsonString() {
 
 void _testGetJsonFromJsonMap() {
   test('Test getJson from JSON map.', () {
-    final json = JsonResponse.fromMap(value: {
+    final json = JsonImpl.fromMap(value: {
       'test1': {"test2": true}
     });
     expect(json.isEmpty, false);
 
-    final childJson = json.getJson(key: 'test1');
+    final childJson = json.get(key: 'test1');
     expect(childJson.isEmpty, false);
     expect(childJson.getBool(key: 'test2'), true);
   });
@@ -205,25 +204,30 @@ void _testGetJsonFromJsonMap() {
 
 void _integrationTest() {
   test('Integration test.', () {
-    final json = JsonResponse.fromString(value: '''
-        {
-            "strength_bars": 1,
-            "infinitive": null,
-            "normalized_string": "Chuang ",
-            "pos": null,
-            "last_practiced_ms": 1573347371000,
-            "skill": "Home 1",
-            "related_lexemes": ["example_a", "example_b"],
-            "test_int_values": [1111, 2222],
-            "last_practiced": "2019-11-10T00:56:11Z",
-            "strength": 0.22,
-            "skill_url_title": "Home-1",
-            "gender": null,
-            "id": "f763b975c30e465d48f3eccbbdd8843a",
-            "lexeme_id": "f763b975c30e465d48f3eccbbdd8843a",
-            "word_string": "窓",
-            "test_bool": true
-         }
+    final json = JsonImpl.fromString(value: '''{
+    "strength_bars": 1,
+    "infinitive": null,
+    "normalized_string": "Chuang ",
+    "pos": null,
+    "last_practiced_ms": 1573347371000,
+    "skill": "Home 1",
+    "related_lexemes": [
+        "example_a",
+        "example_b"
+    ],
+    "test_int_values": [
+        1111,
+        2222
+    ],
+    "last_practiced": "2019-11-10T00:56:11Z",
+    "strength": 0.22,
+    "skill_url_title": "Home-1",
+    "gender": null,
+    "id": "f763b975c30e465d48f3eccbbdd8843a",
+    "lexeme_id": "f763b975c30e465d48f3eccbbdd8843a",
+    "word_string": "窓",
+    "test_bool": true
+}
       ''');
 
     expect(json.isEmpty, false);